toolchain: Replace GCC_VERSION and BUILD_ASSERT macros in gcc.h
GCC_VERSION is defined in a few modules, and those headers are often
included first, so we need to replace that definition with the
Zephyr-specific one.

BUILD_ASSERT is also defined in include/toolchain/common.h, which might
get included before gcc.h. We want to use the gcc-specific one instead
of the general one.

Signed-off-by: Keith Packard <keithp@keithp.com>