[TEST] ci: twister: Run on merge_group
This commit updates the twister workflow to trigger on the
'merge_group' event, which is raised when a pull request is enqueued
into a merge queue.

The 'merge_group' event is handled in the same manner as the 'push'
event because its purpose is to ensure that the changes in the queued
pull request(s) will pass the "main CI" (i.e. 'push' event) before
actually merging the commits into the base branch.